extract theme, use colors from og doom swatch
